requirements.txt
django==5.0.2 
loguru==0.7.2
Pillow==10.0.0
psycopg2==2.9.7
requests==2.31.0
django-4-jet==1.0.9
python-decouple==3.8
djangorestframework==3.14.0 
djangorestframework-simplejwt==5.3.1